requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-amazon==9.9.0
apache-airflow-providers-common-sql==1.27.2
apache-airflow-providers-http==5.3.1
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-snowflake==6.4.0
apache-airflow-providers-standard==1.3.0
boto3==1.38.44
botocore==1.38.44
s3fs==2025.7.0
jsonpath-ng==1.7.0
pandas==2.1.4
numpy==1.26.4
pandera==0.24.0
snowflake-connector-python==3.15.0
snowflake-sqlalchemy==1.7.5
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
SQLAlchemy-JSONField==1.0.2